Update choosing-sender-pubkey.md
Type, 'becuase' should be 'because'. 

But also, I was curious about this repo. Do we still wish to actively document specs here? And if we do wish to more thoroughly document specs, perhaps it should go in flowcrypt-docs?

Of course, there was probably a reason for this to exist as it is, which is what made me curious.